[IMP] web: convert core/utils/render tests to hoot
task-id: 3705027
