--- addons/web/static/tests/legacy/core/utils/render_tests.js
@@ -1,31 +1,0 @@
-/** @odoo-module alias=@web/../tests/core/utils/render_tests default=false */
-
-import { renderToElement, renderToString } from "@web/core/utils/render";
-
-QUnit.module("utils", () => {
-    QUnit.module("render");
-
-    QUnit.test("renderToElement always returns an element", (assert) => {
-        renderToString.app.addTemplate(
-            "test.render.template.1",
-            `<t t-if="False">
-              <div>NotOk</div>
-            </t>
-            <t t-else="">
-              <div>Ok</div>
-            </t>`
-        );
-        const compiledTemplate = renderToElement("test.render.template.1");
-        assert.strictEqual(
-            compiledTemplate.parentElement,
-            null,
-            "compiledTemplate.parentElement must be empty"
-        );
-        assert.strictEqual(
-            compiledTemplate.nodeType,
-            Node.ELEMENT_NODE,
-            "compiledTemplate must be an element"
-        );
-        assert.strictEqual(compiledTemplate.outerHTML, "<div>Ok</div>");
-    });
-});
